Move the UI components in 'components' folder
