+++ pkg/ui/src/components/SVGTest/SVGTest.stories.tsx
