+++ pkg/ui/src/components/SVGTest/__tests__/SVGTest.test.tsx
