+++ pkg/ui/src/index.ts
@@ -1,6 +1,6 @@
-import Button from "./Button";
-import SideNavButton from "./SideNavButton";
-import SVGTest from "./SVGTest";
+import Button from "@/components/Button";
+import SideNavButton from "@/components/SideNavButton";
+import SVGTest from "@/components/SVGTest";
 
 // Main CSS is imported here to signal the bundler the usage of Tailwind
 // in order to generate a PostCSS built css file next to the bundle
